+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Huff Neo Compiler changelog
 
 ## [Unreleased]
+- Throw an error if an argument (e.g. `<arg>`) is used but not defined (Fixes: #46).
 
 ## [1.1.5] - 2025-03-19
 - Support nesting of macro calls e.g. `MACRO1(MACRO2(0x1, 0x2), 0x3)`. (See: #40)